+++ src/effects/spells/enlargeReduce.js
@@ -3,6 +3,11 @@ import DDBMacros from "../DDBMacros.js";
 import { effectModules } from "../effects.js";
 
 export async function enlargeReduceEffect(document) {
+
+  const macroToggle = `<br><p>[[/ddbifunc functionName="enlargeReduce" functionType="spell"]]{Enlarge/Reduce Macro}</div></p>`;
+  document.system.description.value += macroToggle;
+  if (document.system.description.chat !== "") document.system.description.chat += macroToggle;
+
   if (!effectModules().atlInstalled) return document;
 
   let effect = baseSpellEffect(document, document.name);